Hide content until first client render
This should be considered as a temporary hack.

During server side renders, HDS styles can't be applied because there's
no document object. HDS attempts to embed styles into the head of the
document.

This causes a flash of unstyled content, which creates friction during
demos. As a workaround for now, we are postponing the first visible
render of the page to stop the untyled content from being showed.