Removed the HDS utility row from footer
Removed the 2nd line from footer, since we don’t have the feedback feature yet (so the link does not do anything) and “backToTopLabel“ -feature is hard-coded to be in Footer-utilities by HDS. The middle row does not have any tests either. The second row just looked empty and funny as is also described in the Jira ticket.

LIIKUNTA-193