Don't store pagination parameters in search
After the user refreshes, the current pagination logic isn't able to
hydrate earlier pages. This means that instead of a full list of
results, including the current page, only the current page is returned.

After this change the application loses its pagination state on
refresh, which resets the page. The user will see the first page of
results instead of the page that was last loaded into memory.